小重构
