--- .vscode/launch.json
@@ -1,120 +1,0 @@
-{
-    // 使用 IntelliSense 了解相关属性。 
-    // 悬停以查看现有属性的描述。
-    // 欲了解更多信息，请访问: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "lldb",
-            "request": "launch",
-            "name": "Debug unit tests in library 'web'",
-            "cargo": {
-                "args": [
-                    "test",
-                    "--no-run",
-                    "--lib",
-                    "--package=web"
-                ],
-                "filter": {
-                    "name": "web",
-                    "kind": "lib"
-                }
-            },
-            "args": [],
-            "cwd": "${workspaceFolder}"
-        },
-        {
-            "type": "lldb",
-            "request": "launch",
-            "name": "Debug executable 'web'",
-            "cargo": {
-                "args": [
-                    "build",
-                    "--bin=web",
-                    "--package=web"
-                ],
-                "filter": {
-                    "name": "web",
-                    "kind": "bin"
-                }
-            },
-            "args": [],
-            "cwd": "${workspaceFolder}"
-        },
-        {
-            "type": "lldb",
-            "request": "launch",
-            "name": "Debug unit tests in executable 'web'",
-            "cargo": {
-                "args": [
-                    "test",
-                    "--no-run",
-                    "--bin=web",
-                    "--package=web"
-                ],
-                "filter": {
-                    "name": "web",
-                    "kind": "bin"
-                }
-            },
-            "args": [],
-            "cwd": "${workspaceFolder}"
-        },
-        {
-            "type": "lldb",
-            "request": "launch",
-            "name": "Debug unit tests in library 'web_api'",
-            "cargo": {
-                "args": [
-                    "test",
-                    "--no-run",
-                    "--lib",
-                    "--package=web_api"
-                ],
-                "filter": {
-                    "name": "web_api",
-                    "kind": "lib"
-                }
-            },
-            "args": [],
-            "cwd": "${workspaceFolder}"
-        },
-        {
-            "type": "lldb",
-            "request": "launch",
-            "name": "Debug executable 'web_api'",
-            "cargo": {
-                "args": [
-                    "build",
-                    "--bin=web_api",
-                    "--package=web_api"
-                ],
-                "filter": {
-                    "name": "web_api",
-                    "kind": "bin"
-                }
-            },
-            "args": [],
-            "cwd": "${workspaceFolder}"
-        },
-        {
-            "type": "lldb",
-            "request": "launch",
-            "name": "Debug unit tests in executable 'web_api'",
-            "cargo": {
-                "args": [
-                    "test",
-                    "--no-run",
-                    "--bin=web_api",
-                    "--package=web_api"
-                ],
-                "filter": {
-                    "name": "web_api",
-                    "kind": "bin"
-                }
-            },
-            "args": [],
-            "cwd": "${workspaceFolder}"
-        }
-    ]
-}
