--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "rust-analyzer.linkedProjects": [
-        ".\\web\\Cargo.toml",
-    ]
-}
