--- web/src/router/mod.rs
@@ -1,2 +1,0 @@
-mod init_router;
-pub use init_router::*;
