--- web_api/src/api_router/mod.rs
@@ -1,2 +1,0 @@
-mod init_api_routers;
-pub use init_api_routers::*;
