+++ web_dynamic/src/data_struct/upload_user_profile_photo_url.rs
