+++ web_dynamic/src/data_struct/upload_user_profile_photo_url_statement.rs
