+++ web_dynamic/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod api_router;
+pub mod routers;
 pub mod data_struct;
 pub mod database_struct;
