+++ web_dynamic/src/main.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 use warp::Filter;
-use web_api::{api_router::*, database_struct::*};
+use web_api::{routers::*, database_struct::*};
 #[derive(serde::Deserialize, serde::Serialize)]
 struct ProvinceList {
     message: String,
