+++ web_dynamic/src/routers/mod.rs
@@ -1,0 +1,2 @@
+mod init_routers;
+pub use init_routers::*;
