+++ web_static/src/router/mod.rs
@@ -1,0 +1,4 @@
+mod init_router;
+mod init_api_router;
+pub use init_router::*;
+pub use init_api_router::*;
