+++ web_static/src/upload_user_profile_photo/tools.rs
