+++ web_static/src/upload_user_profile_photo/upload_user_profile_photo.rs
