+++ 头像上传测试.html
@@ -25,7 +25,7 @@
             let fileBlob = new Blob([file], { type: 'application/json' });
             formData.append('photo', file);
             try {
-                const response = await fetch('http://127.0.0.1:8080/upload_user_profile_photo', {
+                const response = await fetch('http://127.0.0.1:8080/static/api/upload_user_profile_photo', {
                     method: 'POST',
                     body: formData
                 });
